+++ src/app/analysis/run-analysis/group-analysis/annual-analysis-summary/annual-analysis-summary.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { AnalysisService } from 'src/app/analysis/analysis.service';
-import { AnalysisCalculationsHelperService } from 'src/app/analysis/calculations/analysis-calculations-helper.service';
+import { AnalysisCalculationsService } from 'src/app/analysis/calculations/analysis-calculations.service';
 import { ModifiedEnergyIntensityService } from 'src/app/analysis/calculations/modified-energy-intensity.service';
 import { AnalysisDbService } from 'src/app/indexedDB/analysis-db.service';
 import { FacilitydbService } from 'src/app/indexedDB/facility-db.service';
@@ -20,7 +20,7 @@ export class AnnualAnalysisSummaryComponent implements OnInit {
   facility: IdbFacility;
   annualAnalysisSummary: Array<AnnualAnalysisSummary>
   constructor(private analysisService: AnalysisService, private analysisDbService: AnalysisDbService, private facilityDbService: FacilitydbService,
-    private modifiedEnergyIntensityService: ModifiedEnergyIntensityService, private analysisCalculationsHelperService: AnalysisCalculationsHelperService) {
+    private analysisCalculationsService: AnalysisCalculationsService) {
   }
 
   ngOnInit(): void {
@@ -28,11 +28,7 @@ export class AnnualAnalysisSummaryComponent implements OnInit {
     this.analysisItem = this.analysisDbService.selectedAnalysisItem.getValue();
     this.group = this.analysisService.selectedGroup.getValue();
     this.facility = this.facilityDbService.selectedFacility.getValue();
-    if(this.group.analysisType == 'modifiedEnergyIntensity'){
-      this.annualAnalysisSummary = this.modifiedEnergyIntensityService.getAnnualAnalysisSummary(this.group, this.analysisItem, this.facility);
-    }else{
-      this.annualAnalysisSummary = this.analysisCalculationsHelperService.getAnnualAnalysisSummary(this.analysisItem, this.facility, this.group);
-    }
+    this.annualAnalysisSummary = this.analysisCalculationsService.getAnnualAnalysisSummary(this.analysisItem, this.facility, this.group);
   }
 
   setDataDisplay(display: 'table' | 'graph') {
