+++ src/app/analysis/run-analysis/group-analysis/monthly-analysis-summary/monthly-analysis-summary.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { AnalysisService } from 'src/app/analysis/analysis.service';
-import { AnalysisCalculationsHelperService } from 'src/app/analysis/calculations/analysis-calculations-helper.service';
+import { AnalysisCalculationsService } from 'src/app/analysis/calculations/analysis-calculations.service';
 import { AnalysisDbService } from 'src/app/indexedDB/analysis-db.service';
 import { FacilitydbService } from 'src/app/indexedDB/facility-db.service';
 import { MonthlyAnalysisSummary } from 'src/app/models/analysis';
@@ -20,14 +20,14 @@ export class MonthlyAnalysisSummaryComponent implements OnInit {
   facility: IdbFacility;
   itemsPerPage: number = 12;
   constructor(private analysisService: AnalysisService, private analysisDbService: AnalysisDbService,
-    private analysisCalculationsHelperService: AnalysisCalculationsHelperService, private facilityDbService: FacilitydbService) { }
+    private analysisCalculationsService: AnalysisCalculationsService, private facilityDbService: FacilitydbService) { }
 
   ngOnInit(): void {
     this.dataDisplay = this.analysisService.dataDisplay.getValue();
     this.analysisItem = this.analysisDbService.selectedAnalysisItem.getValue();
     this.group = this.analysisService.selectedGroup.getValue();
     this.facility = this.facilityDbService.selectedFacility.getValue();
-    this.monthlyAnalysisSummary = this.analysisCalculationsHelperService.getMonthlyAnalysisSummary(this.group, this.analysisItem, this.facility);
+    this.monthlyAnalysisSummary = this.analysisCalculationsService.getMonthlyAnalysisSummary(this.group, this.analysisItem, this.facility);
   }
